apps/fleet/pyproject.toml
[project]
name = "projectplant-fleet"
version = "0.1.0"
description = "ProjectPlant fleet control plane"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.2.0",
  "PyNaCl>=1.5.0",
  "python-multipart>=0.0.9",
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["config", "main"]

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
  "black>=24.4.0",
  "mypy>=1.10.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
